Let the multi-fallback handle Any case as well
A good night of sleep revealed the solution: using the +@p signature
instead of the **@p signature on the fallback, should fix the issue
found with Z and X using set operators that don't return a Bool.

This caused one complication: the Any candidates of these set operators
are apparently narrower than the +@p candidates, so these are now
removed and the logic for handling a single value is now handled by
the +@p fallback candidates.