Generate a `rakuw.exe` on Windows
On Windows executables are either marked as "console" applications or
"windows" applications. The first always spawn a console window to which
STDIN, STDOUT and STDERR are connected, the second spawns no such window
and STDIN, STDOUT and STDERR are not connected to anything. This can't be
decided at runtime, but is a flag set on the executable file itself. So the
best we can do to support both usage scenarios is provide separate binaries
for each scenario.
Python does the same. The `python.exe` executable is compiled as "console"
while `pythonw.exe` is a "windows" executable. There is no point in making
up a new naming scheme, so we just do what the others do and name it
`rakudow.exe` / `rakuw.exe`.

Fixes #3582