Make %h{;} slices with * non-assignable
Because a slice with * at any level, is non-deterministic.  Making it
die with an appropriate error message like:

    $ raku -e 'my %h; %h{*} = 42'
    Cannot assign to *, as the order of keys is non-deterministic

would require mixing in some logic in the return List, or replacing
any containers with custom containers that produce that error message.
Which would incur a significant overhead for the case when no assignment
is attempted.