Implement @[;] with all the bells and whistles
Since there are subtle differences with the current (incomplete)
implementation, this is only available in v6.e.PREVIEW.  Allows for
use of * , and WhateverCode such as *-1, as well as iterables at any
level.  Optimised for the simple 2 and 3-dimensional cases, but due
to checks for named parameters, is about 1% slower than the current
implementation.

Allows for all supported adverbs already supported on multi-level hash
access: :exists :!exists :delete :k :kv :p :v and valid combinations
thereof.