Implement binding of hash-slices
AKA %h<a b c> := 1,2,3 .  For some reason, this was not implemented
before, perhaps because of pre-GLR peculiarities making this difficult.
Implementing turned out not to be such a big thing, so here goes.

This breaks 2 spectests that specifically test for throwing an
X::Bind::Slice exception:

  S03-binding/hashes.t, 38
  S32-exceptions/misc2.t, 47

This makes 10 TODO'd spectests pass:

  t/spec/S32-hash/slice.t, 15-22, 25-26