Introducing CachedIterator role
This is a refinement of the PredictiveIterator role, in that it allows
an Iterator class to provide the cache of the Seq, so that Lists and
Arrays that have already been fully reified, do not need to generate
a cache.  Seems like an easy applicable idea, but unfortunately, this
hit some snags.

First of all, apparently it is legal to advance the iterator on a Seq
*before* .cache is called:

    my @a = ^10; dd @a.Seq.skip(5).cache;  # (5, 6, 7, 8, 9)

This makes it impossible to just have the .cache method of a
CachedIterator return the source wrapped in a List.  One would think that
just wrapping a .splice of the source into a List would also work, but
alas, this causes as yet unexplained spectest errors.  So, for now, in
that case we'll just re-instate the original behaviour in that case, which
is List.from-iterator from the iterator.

How much does this save?  An artificial case such as:

    my @a = ^1000; @a.Seq[*-1] for ^10000;

is 14x as fast.  Note that this situation actually occurs for:

    my @a = ^1000; @a.sort[*-1] for ^100;

but then the savings are only about 3%, because they are drowned out
by the effort of sorting.

This currently breaks 4 spectests:

    t/spec/S32-list/seq.t: 28, 38, 42, 44

These all basically test whether an empty Seq will die because it is
exhausted.  Not sure about the validity of these tests.