Use nqp::can to test for test-assertion
In some cases, specifically on the JVM, frames can be NQP code objects.
So instead of using .? dispatch, we're just checking if it can do the
'is-test-assertion' method.  Which it only can if the code actually had
the "is test-assertion` trait applied with a true value.

This should fix #3999, Christian Bartolomäus++ for spotting.