Expose ObjectHash/TypedHash roles
These roles used to be internal to the Hash class.  By exposing them
it is possible to use them for easy typechecking and dispatch.

This makes the rather esoteric check:

    nqp::eqaddr(foo.keyof,Str(Any))

much more readable as:

    nqp::istype(foo,ObjectHash)

albeit with inverted meaning, of course,  This check is about 2x as
fast as the nqp::eqaddr version.  Still it is little work, but some
of these happen in potentially hot code paths in set operator handling.

For easier maintainability, these roles have been put into separate
source-files.

This commit also changes all places where the faster ObjectHash check
could be used, and adds the two new visible roles in the core tests.

One could possibly bikeshed that these roles should be called
Hash::ObjectHash and Hash::TypedHash respectively.  If that is the
consensus, then that could be changed.