Fix for isa-ok
It was assuming that the test subject inherits from `Mu` which is
sometimes incorrect. For example, it fails on

    isa-ok Any.HOW, Metamodel::ClassHOW;

To make it working I added special handling for stringified type names
and replaced the call to method `isa` with `nqp::istype` in all other
cases.

This fix also makes a few TODOs in roast passing due to use of `isa-ok`
to test against roles.