Make method Bridge take care of uninitialized numerics
I needed all this to take care of `Int() == 0` case.

For typeobject it'd throw a new X::Numeric::Uninitialized exception.

For coercions it'd bypass to Mu::Numeric to produce a warning and
use it's return as the default.

Definite values are special-cased by some of the numeric types.

As a side effect this commit results in more user-friendly error
messages.