Gen faster code for some kinds of `when` cases
If it's a type or literal, we don't need to do a full smartmatch.
Unfortunately this isn't dramatically faster, since the primary cost is
in generating a Backtrace for the Failure created when a case doesn't
match.

An example test case:
```raku
my $a;
my @a = 1, 5, Int, "hi", "bye", 50, Complex;
my $s = now;
for ^100_000 -> $i {
    $a = do given @a[$i % 7] {
        when Int  { "Int" }
        when 5    { "five" }
        when "hi" { "hello" }
        default   { "default" }
    }
};
say now - $s;
say $a
```
takes ~1.4s with this change, ~1.8s before.