Fix incorrect assignment passing with coercion
Fix for #4040. `my Rat(Str) $foo = 1` will now fail because `Int` is not
passing the type check.

Also made `Metamodel::CoercionHOW::coerce()` more strict about the value
it's trying to coerce. Now it will throw if it doesn't match against
`$!constraint_type`. Normally this shouldn't happen to a user unless it
invokes the method directly, but should be of great help in catching
core errors.