Fix building on Windows with MinGW
The flag `--subsystem=windows` is possibly supported by the gnu linker, but
it's not by GCC. So we use `-mwindows` which is supported by GCC. It is
possible that flag also automatically makes GCC link some additional
libraries, but that won't hurt.

Also switch to using `wWinMain`, as the gcc toolchain seems to not accept
a `WinMain` (possibly because of the `-municode` flag). The MSVC toolchain
doesn't seem to mind.

Fixes #4009