Add replacement mechanizm for P6EX HLL symbol
`Perl6::Metamodel::Configuration` can now be told about the top-level X
package where it later looks up for exceptions.

It also got a new method `throw_or_die` which means we first try to find
and throw the requested exception. If it's not found then `nqp::die`
with an alternate message provided.

The first two parameters of the method are exception name and the
alternative message. The rest are passed to the exception constructor as
is:

    Perl6::Metamodel::Configuration.throw_or_die(
        'X::TypeCheck::Binding',
        'This message we use for nqp::die',
        :$got, :$expected, :$symbol, :$parameter, :$constraint
    )

The first usage is comming in a next commit with
`Perl6::Metamodel::LanguageRevision`.