Add "deterministic" info to many iterators
A recap: when an iterator.deterministic call returns False, it means
that the iterator may not produce the same values, or not in the same
order, given the same initial condition.  This specifically applies
to iterators running on hashes at some level, and iterators that create
a random selection on a set of values.

This commit also covers iterators that run off of other iterators.
Although the more complicated ones are not covered in this commit.