Make the REPL a little bit smarter
Basically, any line that does *not* create any direct output to
STDOUT, becomes part of the code that will be executed whenever
the next line is input.  This is a behaviour similar to what is
done when a line is deemed incomplete.  The only difference is,
is that this new behaviour will not be visible because of a prompt
change.

This fixes a number of issues with information being lost between
lines, makes a TODO pass in t/02-rakudo/repl.t and is basically a
fix for issues #2284 and #3925.  Matias Linares++ Altai-man++ for
keeping me tortured :-)

Please be aware that this potentially opens a whole new nest of
REPL bugs.  We'll cross that bridge when we get there.