Implement metamodel transparency of nominalizables
As a matter of fact, a nominalizable is a wrapper around a nominal or
another nominalizable. As such, it makes sense to have their MOP classes
transparent, similar to how their typeobjects are transparent down to
the final nominal type. I.e., it means that for

    role R[::T] {
        my subset RS of T;
        method foo(RS() $v) {...}
    }

`RS().HOW.archetypes.generic` should be true because `RS` wraps a
generic. Similarly, `RS` would have it's `coercive` archetype set would
it be declared as `my subset RS of T()`. The transparency lets the
compiler and runtime to make sure no nominalization is missed however
they're nested into each other. In particular, it makes the sample above
possible and it works with this commit.

Introducing a new metmodel role `Perl6::Metamodel::Nominalizable` which
provides basic common functionality of nominalizables. In particular,
it implements `.^wrappee` method which allows to fetch necessary
nominalization from the nesting. `RS.^wrappee(:coercion)` would return
`T()` for the latter `RS` example. Correspondingly,
`RS().^wrappee(:subset)` would result in `RS` itself.

Implemented `instantiate_generic` for subsets.

Implemented support of the new transparency in `Actions.nqp`. This last
part is a draft for now because the old support of typecaptures in the
signatures stopped working. The new implementation requires
`instantiate_generic` to be invoked on parameter type due to the changes
of it being not the original type capture, but a nominlazation wrapping
it. The current draft implementation uses `curlexpad` for this. But this
would be another thing breaking inlining (on top of the already existing
one). But it should be possible to replace it with a dynamically built
hash of currently known type captures.