Add :emit-timed to Supply.batch(:$seconds)
When :emit-timed is specified, a timer will be running emitting
whatever is in the batch every given number of seconds *if* there
is something in the batch.  The default is to only emit if a new
value is received and it arrives in a new period.

This removes the previous proposal for :emit-on-empty and :emit-once-on-empty.