Optimze binding code by removing a method call
Replace a call to `coercive` method of Parameter with `getattr` and
`bitand_i`. This brings the overall rakudo performance almost on a par
with v2020.10 with euler script, as mentioned in #4056. In my tests the
script is just 3% slower with `time raku euler-10.p6`.

The whitenoise script, mentioned in the same ticket, is 55% faster than
PR #4059, and 2.1x times faster than v2020.11. Unfortunately, it's still
4.5 times slower than with v2020.10. But I hope that inlinable
`nqp::curcode` would resolve this too.