Make any STDERR output also reset REPL
Fixes #4077.  Any output on STDERR will now also cause a "reset" in
the sense that it won't combine the previous lines into a single
EVAL, as that most likely indicated a "worry" warning (such as
described in the issue) or some explicit output on STDERR (such as
a "note").

This means that the redeclaration warning will be shown once.  Which
I think would actually be a nice additional feature.