Don't use nqp::iterator iterating over mutable QuantHashes
This used to depends on R:I.ITERATIONSET2LISTITER that would create
a native string array of the keys of a QuantHash and then return an
iterator for it.  Changed it to return the native string array instead
and adapted the users of it.  Also changed the name of the internal
method to a more descriptive: IterationSet2keys

Also other assorted cleanup in this area to make better readable code.