Add support for native slice assignment from native
Now that slices on native arrays return arrays of the same native
type, it makes sense to optimize assigning those values to a native
array.  This commit makes @a[1,2,3] = @b where @a and @b are of the
same native type, 26x as fast.  And makes it 18x as fast if the
receiving array is a native shaped array.