Make Iterator.sink-all default to using skip-one
This should make no change if an Iterator:
- already has a sink-all method
- does not have its own .skip-one method

This will make a difference if an Iterator has its own .skip-one.
But the premise is that *if* an iterator has its own .skip-one,
it means that the Iterator itself is either side-effect free, *or*
the .skip-one method also handles the side-effects.

This should make a .sink-all on an Iterator that is side-effect free
and has its own .skip-one *significantly* faster.