Remove return Sig from .elems methods
We can be pretty sure that these will either return an Int or a
Failure.  So there's no real need to check here.  This appears to
have a positive effect, typically on multi-dimensional shaped array
accesses, hence the reason for doing this (typically from 15% of
total to 15% of total effort).