Use `exec` instead of `system` and then exiting with the return code.
The approach using `system` misses the case where the invoked program exits
with a signal - for example if the OOM killer takes out the setting build.

Before this change the wrapper script would calmly take the value 137
returned from `system`, shift it right by 8 and exit with status 0, which
make took as success and carried on. At which point the next command invoked
would mysteriously fail due to a missing input file.

Now we see:
    The following step can take a long time, please be patient.
    Stage start      :   0.001
    Stage parse      : 745.469
    Stage syntaxcheck:   0.000
    Stage ast        :   0.000
    Stage optimize   : 102.057
    Stage mast       : Killed
    make: *** [Makefile:1170: blib/CORE.c.setting.moarvm] Error 137

a useful error message and the build aborts.

Where we still need `system` in Configure.pl, simply check whether it
returned non-zero. This way we also catch the error case of death by fatal
signal.

I checked perlport - `exec` and `system` have identical portability
caveats, so this change *should* not cause problems (at least, not cause
problems that we didn't already expose ourselves to).