NativeCall workaround for slowdown of calls to routines with generic arguments
If a routine's signature contains any generics (::T), we generate code to check
at runtime, if the passed type is coercive. This is very expensive and costs
some 70 % of performance in some benchmarks.
For NativeCall there is a workaround though: the only generic parameter is
::?CLASS in the signature. Since the typed CArray roles are only known within
the CArray class, we can be sure that ::?CLASS is actually CArray or maybe a
subclass. Thus we can make the type constraint more concrete.
Brings whitenoise.p6 backup up from 45ish fps to > 180 fps