Re-imagine array slicing once again (now 5x as fast)
As vrurg++ suggested, it makes sense to create a role with the
necessary attributes that will actually do the work, using a private
method that needs to be supplied by the consuming class.

That works, but because calling a private method of a consuming class
from code in the role, is actually currently not optimized.  So this
would lose all of the benefits of this approach.

But what is a class consuming a role different from "copying" the code
of the role.  Not a lot.  So instead of creating a role, this commit
creates a helper script that generates the necessary classes (all 58
of them).  Then these classes are put into dispatch tables (one for the
lazy positions case, and one for the non-lazy positions case), and
the actual slicing now becomes:

- lookup dispatch index
- convert dispatch index to class
- create the object for the iterable
- run the slice logic and return its result

The result of this is that this approach is about 60% faster than
before, bringing the total speedup to about 5x as fast.