Re-imagine @a[] (zen-slice) to new infrastructure
This moves the handling of zen-slices and its adverbs to the new
infrastructure, making the bare zen-slice 12x as fast, and a zen-slice
with adverbs about 6x as fast.

This breaks a small number of tests in roast because the error message
on disallowed adverb combinations, has become more specific (e.g.
"zen slice" instead of "slice").

This also removes the now obsolete ZEN-POS method, which was an
implementation detail.