Remove a lot of unnecessary int $pos candidates
of (AT|ASSIGN|BIND|EXISTS|DELETE)-POS candidates.  These candidates
were once written with the assumption that using a native integer
on a candidate with a native int "receiver", would be more efficient.
It currently isn't, it's even (much) slower than having a native
integer on an Int "receiver".

Also, turned out that Array.DELETE-POS(Int) was deferring to the
Array.DELETE-POS(int) candidate.  By making that the Int candidate
and removing the int candidate, DELETE-POS is now about 7% faster.