Introducing Seq.slice(indices) (#3983)
Introducing Seq.slice(indices)

It appears that a "slice" method on Seq has been discussed extensively
in the longer ago past.  So I thought I'd have a go at it today.

The .slice method takes any number of indices that are monotonically
increasing (that is, any later index value *must* be greater than any
previous index value), and produces the associated values from the
iterator of the given sequence.  To give an artificial example:

    my @a = ^10; dd @a.Seq.slice(0, 2 ... *);  # (0, 2, 4, 6, 8).Seq

is about 2x as fast as:

    my @a = ^10; dd @a.Seq[0, 2 ... *];       # (0, 2, 4, 6, 8)

This is because the .slice method does *not* have to do any caching.

Note that if an index is produced that does not follow the rule, an
exception will be thrown.

Turns out the underlying iterator had already been implemented for some
aspects of Match processing, so this was actually only creating a frontend
for that functionality.