Implement `sec`, `asec` and `sech` in terms of other NQP trigonometric funcs.
There is some inconsistency in MoarVM - as well as the usual trigonometric
functions `sin`, `cos`, `tan` and their inverses, *some* of the others are
implemented directly as NQP ops, but not all. So we have nqp::sec and
nqp::asec for secants, but no OPs for cosecants or cotangents (or their
inverses), with the result that these are implemented by Num.pm6 as
arithmetic relations using the OPs that exist.

Similarly, `sech` exists as a wrapper, but not `cosech` or `cotanh` (or any
of the inverses).

Looking at MoarVM's interp.c I see that the 3 OPs in question *don't* call
C library functions for the primitives - each is C code implementing the
operation in terms of arithmetic on other library functions - library
functions that we expose as NQP ops.

Hence it's much simpler if we move this arithmetic from MoarVM to the CORE:
1) We can remove the OPs from MoarVM
2) The code can now be JITted (everything that wrapper uses is in the JIT,
   whereas the 3 synthesised functions are not)
3) The 3 OPs don't seem to be implemented in the JVM or JS backends, so we
   also improve their test pass percentage :-)