Add candidate for native slice assignment from Iterable
Makes slice assignment on a native array from an Iterable
about 15x as fast.  Also makes it more correct:

If the values for a slice assignment to a native array are not of
the same type, it would fall back to the generic slice assignment.
This would a. return a List rather than a native array of the same
type, and b. would not die on passing nested indices like with other
slices on native arrays.