Re-imagine array slicing, fixing issues, between 2x and 10x faster (#4131)
This commit makes array slice accesses and assignment between 2x 
and 10x as fast *and* fixes several issues found in the previous
implementation.  It also creates a framework for handling adverbs that
are allowed with postcircumfix adverbs (in a separate source-file Rakudo/Internals/PostcircumfixAdverbs).

Specific fixes are with Iterables within Iterables as index.  For example,

  dd <a b c d e>[0,(1,2,3),4] # ("a", ("b", "c", "d"), "e")

worked, but:

  dd <a b c d e>[0,(1,2,3),4]:k # (0,4)

did not, simply ignoring the inner iterable (1,2,3).  This is now
fixed.  This was apparently not implemented correctly since its initial
implementation.

This includes about 6K lines of generated classes.  Once the performance
penalty of calling a private method of a consuming class from within a role
is fixed, this can be greatly reduced by generating roles and consuming
classes.