Streamline some iterator setups
You can assign attributes in a signature like:

    method foo($!foo,$!bar)

but this costs 2 Scalar allocations for each invocation.  In normal
code, one would use "is built(:bind)".  But sadly, this is not possible
at this location in the setting, so "manually" bind the attributes
for these iterators.

This makes array slices such as %h<a b c> about 40% faster.