Fix potential deadlock on specialization of roles
NQPLock's protect method doesn't cope well with return statements. Doing a
return from a protected block will leave the lock in place. The only reason why
this hasn't surfaced as deadlock before is that the lock is using a reeantrant
mutex and most of the time the same thread will do the specializations. Fix by
simply not using return and instead move the remainder into an if statement.
Fixes a hang of role R { }; (^100).hyper.map({R.HOW.specialize(R, Any)})