Introduce Rakudo::Internals.RERUN-WITH method
Takes a named parameter indicating which environment variable should
be set to which value.  If this environment variable was not set yet,
will call the current program with all its parameters again and return
a Proc::ASsync object for that.  If the environment variable was already
set, then it will just retirn NIL.

This should allow for some internal modules to process the output of
a program run with such an environment variable set, such as with
the MoarVM::SIL module.

Note that once the handling of the Proc::Async object is complete, it
is expected that the process is exited immediately, to prevent the code
actually being run twice.

This also moves around some stubs to allow for earlier reference to
Proc::Async in the setting.

This also changes the recently introduced Rakudo::Internals.E method:
changed it to PROGRAM as it now reflects what the program is to be
executed, whether that be $*PROGRAM-NAME or whatever was specified
with -e.