Introduce the SIL core development helper class
SIL (short for Spesh Inline Log) is intended to be called with the
-M parameter on the command line to generate a log of the successful
and unsuccessful inlines that spesh has done during execution of the
program that is specified by the rest of the parameters.

This also turns MoarVM::SIL into a proper module that exports the
SIL subroutine.  This subroutine either returns a MoarVM::SIL object
or Nil (when being run in logging mode).  This should now also allow
writing tests to make sure certain subs / methods continue to be inlined
after changes having been made to the core.

A typical test setup would be:

    use Test;
    use MoarVM::SIL;

    if SIL() -> $SIL {
        LEAVE $SIL.exit;

        for <ASSIGN-POS ASSIGN_POS_SLOW_PATH> {
            ok $SIL.inlined-by-name($_), "Was $_ inlined?";
        }
    }

    # the code to run
    for ^100000 {
        my @a; @a[3] = 42;
    }