Alternate fix for JVM breakage to PR #4189
Instead of coercing explicitely, revert to the original (slower)
version for the JVM, as that was a. guaranteed to work, and b.
should allow for easier HLL-fication of the code once we don't
need the nqp ops for performance anymore.  Bartolin++ for the PR
and generally monitoring the JVM backend.