Fix segfault caused by repossession of native call site
We deserialize a native call site. Then we call the function which triggers the
actual setup, i.e. resolving the library name, loading the library, generating
caller code. We run the function often enough that spesh decides its worth its
time. While spesh is working however, we deserialize the module again and as
part of the repossession process clear the native call site's body. I.e.
body->entry_point becomes NULL. This causes us to compile a jump to 0.

The fix is to avoid repossession of the native call site in the first place.
Since the native call site is an attribute of the routine (via the Native role)
and attributes are flattened into the objects holding them, we have to mark the
entire routine neverrepossess.