Introduce the SL core development helper class
SL (short for Spesh Log) is intended to be called with the -M parameter
on the command line to generate a report of the spesh log of the program
that is specified by the rest of the parameters.

This also adds / installs MoarVM::Spesh (an programmatic interface to the
spesh log) and MoarVM::SL (a helper module to run the program with spesh
logging enabled and return an appropriated MoarVM::Spesh object).