Add support for whitespace-separated CLI options
This commit resolves the exiting TODO item:
> Allow both = and space before argument of double-dash args

More specifically, before this commit, a user could pass a `foo`
command that took a named `bar` values in nine ways:

    $ foo --bar=value
    $ foo -bar=value
    $ foo :bar=value
    $ foo --bar      # equivalent to --bar=True
    $ foo -bar
    $ foo :bar
    $ foo --/bar     # equivalent to --bar=False
    $ foo -/bar
    $ foo :bar

(The ways that involve using a single - do not work when bundling is
turned on, since that would be ambiguous.)

This commit adds three more ways:

    $ foo --bar value
    $ foo -bar value
    $ foo :bar value

(Note that this does not limit space separated values to the `--bar`
form.  I'd be happy to add an extra check if anyone would like it, but
I didn't see a reason to add it)

Supporting space-separated options raises one common difficulty.  If
`foo` accepts both an option and a positional argument, then the
following expression is initially ambiguous:

    $ foo --bar baz

Is that a option :bar("baz") and no positional argument?  Or is it
:bar(True) and a positional argument of "baz"?  This commit resolves
that ambiguity by treating all values after an option as values for
that option _unless_ they start with a - or : or if the --bar is type
constrained to be a Bool (which indicates that it isn't an option at
all, but a flag.)  Thus, if foo has the signature

    MAIN(:$bar, Bool :$v, $pos)

It will be parsed as follows:

   $ foo --bar baz     # :bar("baz")
   $ foo --bar -v baz  # :bar(True) :v(True) $pos = "baz"

If a user wants to pass a positional argument after an option, they
can use -- to end parsing of non-positional arguments (which was
supported prior to this commit).