Add BB.specialized attribute, fix name checking
Apparently a BB can be inlined *without* it being specializd, and
that is reported in the log.  Unfortunately, this wasn't taken into
account when looking for names, so name checks (like the
t/09-moar/*inline*.t tests) would report failure, when in fact the
name checked for *was* inlined, but without being specialized.

nine++ for tracking the source of these flappers