Fix for #4216
It turned out that a lazy iterable as part of a non lazy list of indexes
in a slice, had the possibility of hanging because eager semantics where
being used on the lazy iterator, e.g. in `(1..10)[1,^Inf]`

This commit removes all of the Array::Slice classes that existed for
handling lazy indexes.  The candidates for handling lazy indices are
now incorporated in the non-lazy class.

The semantics are now:
- when the iterator of indexes is lazy, then it will stop creating the
  slice as soon as an element is accessed that does not exist
- when an Iterable is found as an index, then a sublist will be created
- if the iterator of a sublist is lazy, it will stop filling the *sublist*
  as soon as an element is accessed that does not exist.

Some examples:

    dd (1..5)[1,^Inf];           # (2, (1, 2, 3, 4, 5))
    dd (1..5)[1,^Inf,3,4];       # (2, (1, 2, 3, 4, 5), 4, 5)
    dd (1..5)[1..10];            # (2, 3, 4, 5, Nil, Nil, Nil, Nil, Nil, Nil)
    dd (1..5)[lazy 1..10];       # (2, 3, 4, 5)
    dd (1..5)[lazy 4..6, 1,2];   # ((5, Nil, Nil), 2, 3)
    dd (1..5)[(lazy 4..6), 1,2]; # ((5,), 2, 3)

This breaks one questionable test in roast: S09-subscript/slice.t #33